fix contract - integration callback working
