+++ contract/test.sh
@@ -1,0 +1,1 @@
+cargo +nightly test
